# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Obiren Production Monolith API",
    "version": "1.0.0",
    "description": "Production API specification for Obiren Women Health, Safety, and Clinical Ecosystem"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local Development Server"
    }
  ],
  "paths": {
...
    "/safety/sos/trigger": {
      "post": {
        "summary": "Trigger Web SOS panic alert",
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/safety/sos/cancel": {
      "post": {
        "summary": "Cancel active SOS with Safety PIN",
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}